requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.28.1
python-dotenv==1.0.0
openai==0.27.8
celery[redis]
pyjwt
requests
openai
kubernetes
redis
pyjwt
cryptography
requests
pyjwt
celery[redis]
fastapi==0.95.1
uvicorn[standard]==0.22.0
kubernetes==26.1.0
cpufeature==0.9.0
torch==2.0.1
petals==0.1
Flask==2.2.5
Flask-Cors==6.0.0
Flask-SQLAlchemy==3.0.0
Flask-RESTful==0.3.9
Flask-Migrate==3.1.0
Flask-WTF==1.0.1
gunicorn==20.1.0
requests==2.31.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
tensorflow==2.11.1
torch==2.0.1
torchvision==0.15.2
transformers==4.28.1
spacy==3.5.0
nltk==3.7
python-dotenv==0.21.1
celery==5.2.6
redis==4.4.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Admin==1.5.8
flask-caching==1.10.1
Flask-Session==0.4.0
flask-socketio==5.3.0
flake8==5.0.4
chardet==5.1.0
isort==5.11.4
codecov==2.1.13
pytest-cov==4.1.0
build==0.10.0
wheel==0.42.0
black==23.3.0
mypy==1.12.0
pytest==7.4.4
pytest-pyodide==0.54.0
pyodide-build==0.24.1
webdriver-manager==4.0.1
keras==2.12.0
imgaug==0.4.0
nlpaug==1.1.3
tsaug==0.2.1
boto3==1.18.69
sagemaker==2.59.3
Pillow==9.5.0
sentencepiece==0.1.98
matplotlib==3.7.1
onnxruntime==1.14.1
librosa==0.9.2
scipy==1.10.1
aiohttp==3.8.4
google-cloud-speech==2.22.2
fastapi==0.95.0
uvicorn==0.22.0
httpx
openai  # or llama-cpp-python, or use LangChain
